requirements.txt
aiodns==3.0.0
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
Brotli==1.0.9
cachetools==4.2.4
cchardet==2.1.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
emoji==2.1.0
frozenlist==1.3.1
google-api-core==1.33.2
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.48.2
idna==3.4
multidict==6.0.2
protobuf==3.20.3
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
yarl==1.8.1
